requirements.txt
numpy>=1.21.6
scikit-image==0.19.0
scikit-learn==1.0.1
#importlib_metadata==4.11.1
#napari[all]==0.4.9 # fixed
#jupyterlab==3.0.13
#ipywidgets==7.6.5
#albumentations==1.1.0
pytorch-lightning==1.4.9 # deliberately choosing < 1.5
#neptune-client==0.14.3
#neptune-contrib==0.28.1
#vispy==0.11.0
certifi==2019.11.28
Click==8.0
Flask
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==3.0
MarkupSafe==2.0.0
Werkzeug==2.0
pickle



tpot>=0.11.0
cloudpickle>=0.6.0
joblib==0.14.0

# Base ------------------------------------------------------------------------
gitpython>=3.1.30
matplotlib>=3.3


opencv-python>=4.1.1
Pillow>=7.1.2
psutil  # system resources
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
thop>=0.1.1  # FLOPs computation
torch>=1.7.0  # see https://pytorch.org/get-started/locally (recommended)
torchvision>=0.8.1
tqdm>=4.64.0
pandas>=1.1.4
seaborn>=0.11.0